Preprocessing script for Mapillary.
This also fixes a bug in the soft resizing that shows that cv2.resize doesn't work with
INTER_AREA if the downsample factor isn't divisible by 2 in both height and width.
This means that my mapillary downsampling is a little different than my older downsampled
CityScapes labels, but I don't think it has significant effect on the results.