Bigger commit adding graph freezing, and 2 ros nodes, for publishing images and segmenting them.
The freeze_graph.py script can be used to create a fixed frozen graph of a training checkpoint.
Using the debugging image publisher random images can be published in a loop.
Finally, running the semantic segmentation node the published images can be segmented.
Based on only needing a frozen graph the dependencies of the node are kept to a minimum.