Added "auto_resume" and "multi location" loading.
The auto resume flag works like the resume flag but will also
work when no data is available just yet. This is useful when
training on a cluster where jobs might be canceled and then
rescheduled later on. The current implementation was a quick
hack and should be refactored at some point.

The multi location loading now allows certain parts of paths to
be in square brackets e.g. /path/to/[option1,option2]/data in
so that an SDD could for example be used when available but
it uses other options when not.